+++ pipelex/libraries/llm_integrations/vertexai.toml
@@ -21,12 +21,28 @@ max_prompt_images = 3000
 cost_per_million_tokens_usd = { input = 0.1, output = 0.4 }
 platform_llm_id = { vertexai = "google/gemini-2.0-flash" }
 
-[gemini."gemini-2.5-pro".latest]
+[gemini."gemini-2.5-pro"."latest"]
 is_gen_object_supported = true
 is_vision_supported = true
 max_prompt_images = 3000
 cost_per_million_tokens_usd = { input = 0.0, output = 0.0 }
 platform_llm_id = { vertexai = "google/gemini-2.5-pro-preview-05-06" }
+
+# Update commented because the latest version is not yet on VertexAI
+
+# [gemini."gemini-2.5-pro"."2025-05-06"]
+# is_gen_object_supported = true
+# is_vision_supported = true
+# max_prompt_images = 3000
+# cost_per_million_tokens_usd = { input = 0.0, output = 0.0 }
+# platform_llm_id = { vertexai = "google/gemini-2.5-pro-preview-05-06" }
+
+# [gemini."gemini-2.5-pro".latest]
+# is_gen_object_supported = true
+# is_vision_supported = true
+# max_prompt_images = 3000
+# cost_per_million_tokens_usd = { input = 0.0, output = 0.0 }
+# platform_llm_id = { vertexai = "google/gemini-2.5-pro-preview-06-05" }
 
 [gemini."gemini-2.5-flash"."2025-04-17"]
 is_gen_object_supported = true
