+++ tests/pipelex/cogt_asynch/test_image_gen.py
