+++ tests/pipelex/cogt_asynch/test_llm_gen_object.py
