+++ tests/pipelex/cogt_asynch/test_llm_vision.py
