+++ tests/pipelex/cogt_asynch/test_ocr.py
